MATERI XII SEMESTER 1/01. Dasar Pemrograman Jaringan/Latihan/sistem_parkir/server/requirements.txt
mysql-connector-python==8.0.33
opencv-python==4.7.0.72
numpy==1.24.3
python-dotenv==1.0.0
